+++ lib/src/shared/widgets/measure_size.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
+
+class MeasureSize extends StatefulWidget {
+  const MeasureSize({super.key, required this.onChange, required this.child});
+
+  final Widget child;
+  final ValueChanged<Size> onChange;
+
+  @override
+  State<MeasureSize> createState() => _MeasureSizeState();
+}
+
+class _MeasureSizeState extends State<MeasureSize> {
+  Size? _oldSize;
+
+  @override
+  Widget build(BuildContext context) {
+    SchedulerBinding.instance.addPostFrameCallback((_) {
+      if (!mounted) return;
+      final size = context.size;
+      if (size != null && size != _oldSize) {
+        _oldSize = size;
+        widget.onChange(size);
+      }
+    });
+    return widget.child;
+  }
+}